Refactor units provider module structure
- Moved `UnitsDisplayName` and `UnitsEssentials` to their respective submodules: `display_name` and `essentials`.
- Updated references throughout the codebase to reflect the new module paths.
- Introduced new files for `UnitsDisplayName` and `UnitsEssentials` with appropriate struct definitions and data markers.
- Adjusted data provider implementations to align with the new structure.

This change enhances organization and clarity within the units provider module.
