+++ components/experimental/src/dimension/provider/units/display_name.rs
@@ -61,7 +61,7 @@ impl databake::Bake for UnitsDisplayName<'_> {
         let bytes = self.patterns.elements.as_bytes().bake(ctx);
         // Safety: The bytes are returned by `PluralElementsPackedULE::slice_as_bytes`.
         databake::quote! { unsafe {
-            icu_experimental::dimension::provider::units::UnitsDisplayName::from_bytes_unchecked(#bytes)
+            icu_experimental::dimension::provider::units::display_name::UnitsDisplayName::from_bytes_unchecked(#bytes)
         }}
     }
 }
