+++ components/experimental/src/dimension/provider/units/mod.rs
@@ -2,6 +2,6 @@
 // called LICENSE at the top level of the ICU4X source tree
 // (online at: https://github.com/unicode-org/icu4x/blob/main/LICENSE ).
 
-pub mod currency;
-pub mod percent;
-pub mod units;
+pub mod display_name;
+pub mod essentials;
+pub mod pattern_key;
