+++ components/experimental/src/lib.rs
@@ -107,8 +107,8 @@ pub mod provider {
         super::dimension::provider::currency::patterns::CurrencyPatternsDataV1::INFO,
         super::dimension::provider::currency::extended::CurrencyExtendedDataV1::INFO,
         super::dimension::provider::percent::PercentEssentialsV1::INFO,
-        super::dimension::provider::units_essentials::UnitsEssentialsV1::INFO,
-        super::dimension::provider::units::UnitsDisplayNameV1::INFO,
+        super::dimension::provider::units::essentials::UnitsEssentialsV1::INFO,
+        super::dimension::provider::units::display_name::UnitsDisplayNameV1::INFO,
         super::displaynames::provider::LanguageDisplayNamesV1::INFO,
         super::duration::provider::DigitalDurationDataV1::INFO,
         super::displaynames::provider::LocaleDisplayNamesV1::INFO,
