+++ provider/data/experimental/stubdata/units_display_name_v1.rs.data
@@ -13,16 +13,16 @@ macro_rules! __impl_units_display_name_v1 {
         #[clippy::msrv = "1.82"]
         const _: () = <$provider>::MUST_USE_MAKE_PROVIDER_MACRO;
         #[clippy::msrv = "1.82"]
-        impl icu_provider::DataProvider<icu::experimental::dimension::provider::units::UnitsDisplayNameV1> for $provider {
-            fn load(&self, req: icu_provider::DataRequest) -> Result<icu_provider::DataResponse<icu::experimental::dimension::provider::units::UnitsDisplayNameV1>, icu_provider::DataError> {
-                Err(icu_provider::DataErrorKind::IdentifierNotFound.with_req(<icu::experimental::dimension::provider::units::UnitsDisplayNameV1 as icu_provider::DataMarker>::INFO, req))
+        impl icu_provider::DataProvider<icu::experimental::dimension::provider::units::display_name::UnitsDisplayNameV1> for $provider {
+            fn load(&self, req: icu_provider::DataRequest) -> Result<icu_provider::DataResponse<icu::experimental::dimension::provider::units::display_name::UnitsDisplayNameV1>, icu_provider::DataError> {
+                Err(icu_provider::DataErrorKind::IdentifierNotFound.with_req(<icu::experimental::dimension::provider::units::display_name::UnitsDisplayNameV1 as icu_provider::DataMarker>::INFO, req))
             }
         }
     };
     ($ provider : ty , ITER) => {
         __impl_units_display_name_v1!($provider);
         #[clippy::msrv = "1.82"]
-        impl icu_provider::IterableDataProvider<icu::experimental::dimension::provider::units::UnitsDisplayNameV1> for $provider {
+        impl icu_provider::IterableDataProvider<icu::experimental::dimension::provider::units::display_name::UnitsDisplayNameV1> for $provider {
             fn iter_ids(&self) -> Result<std::collections::BTreeSet<icu_provider::DataIdentifierCow<'static>>, icu_provider::DataError> {
                 Ok(Default::default())
             }
@@ -31,22 +31,22 @@ macro_rules! __impl_units_display_name_v1 {
     ($ provider : ty , DRY) => {
         __impl_units_display_name_v1!($provider);
         #[clippy::msrv = "1.82"]
-        impl icu_provider::DryDataProvider<icu::experimental::dimension::provider::units::UnitsDisplayNameV1> for $provider {
+        impl icu_provider::DryDataProvider<icu::experimental::dimension::provider::units::display_name::UnitsDisplayNameV1> for $provider {
             fn dry_load(&self, req: icu_provider::DataRequest) -> Result<icu_provider::DataResponseMetadata, icu_provider::DataError> {
-                Err(icu_provider::DataErrorKind::IdentifierNotFound.with_req(<icu::experimental::dimension::provider::units::UnitsDisplayNameV1 as icu_provider::DataMarker>::INFO, req))
+                Err(icu_provider::DataErrorKind::IdentifierNotFound.with_req(<icu::experimental::dimension::provider::units::display_name::UnitsDisplayNameV1 as icu_provider::DataMarker>::INFO, req))
             }
         }
     };
     ($ provider : ty , DRY , ITER) => {
         __impl_units_display_name_v1!($provider);
         #[clippy::msrv = "1.82"]
-        impl icu_provider::DryDataProvider<icu::experimental::dimension::provider::units::UnitsDisplayNameV1> for $provider {
+        impl icu_provider::DryDataProvider<icu::experimental::dimension::provider::units::display_name::UnitsDisplayNameV1> for $provider {
             fn dry_load(&self, req: icu_provider::DataRequest) -> Result<icu_provider::DataResponseMetadata, icu_provider::DataError> {
-                Err(icu_provider::DataErrorKind::IdentifierNotFound.with_req(<icu::experimental::dimension::provider::units::UnitsDisplayNameV1 as icu_provider::DataMarker>::INFO, req))
+                Err(icu_provider::DataErrorKind::IdentifierNotFound.with_req(<icu::experimental::dimension::provider::units::display_name::UnitsDisplayNameV1 as icu_provider::DataMarker>::INFO, req))
             }
         }
         #[clippy::msrv = "1.82"]
-        impl icu_provider::IterableDataProvider<icu::experimental::dimension::provider::units::UnitsDisplayNameV1> for $provider {
+        impl icu_provider::IterableDataProvider<icu::experimental::dimension::provider::units::display_name::UnitsDisplayNameV1> for $provider {
             fn iter_ids(&self) -> Result<std::collections::BTreeSet<icu_provider::DataIdentifierCow<'static>>, icu_provider::DataError> {
                 Ok(Default::default())
             }
