+++ provider/registry/src/lib.rs
@@ -258,8 +258,8 @@ macro_rules! registry(
             icu::experimental::dimension::provider::currency::patterns::CurrencyPatternsDataV1: CurrencyPatternsDataV1,
             icu::experimental::dimension::provider::currency::extended::CurrencyExtendedDataV1: CurrencyExtendedDataV1,
             icu::experimental::dimension::provider::percent::PercentEssentialsV1: PercentEssentialsV1,
-            icu::experimental::dimension::provider::units::UnitsDisplayNameV1: UnitsDisplayNameV1,
-            icu::experimental::dimension::provider::units_essentials::UnitsEssentialsV1: UnitsEssentialsV1,
+            icu::experimental::dimension::provider::units::display_name::UnitsDisplayNameV1: UnitsDisplayNameV1,
+            icu::experimental::dimension::provider::units::essentials::UnitsEssentialsV1: UnitsEssentialsV1,
             icu::experimental::duration::provider::DigitalDurationDataV1: DigitalDurationDataV1,
             icu::experimental::displaynames::provider::RegionDisplayNamesV1: RegionDisplayNamesV1,
             icu::experimental::displaynames::provider::LanguageDisplayNamesV1: LanguageDisplayNamesV1,
