+++ provider/source/src/units/mod.rs
@@ -2,7 +2,7 @@
 // called LICENSE at the top level of the ICU4X source tree
 // (online at: https://github.com/unicode-org/icu4x/blob/main/LICENSE ).
 
-pub(crate) mod data;
+pub(crate) mod display_name;
 pub(crate) mod essentials;
 pub(crate) mod helpers;
 pub(crate) mod ids;
